Move triangular calibration into its own object
